chore(team): continue backend manager splits
